Extract middleToken & isUsernameValid functions to validString.js & middleware.js + Added a body content verification for PUT & PATCH notes endpoint
