+++ app.js
@@ -4,11 +4,9 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
-var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
-var notesRouter = require('./routes/notes');
-var signupRouter = require('./routes/signup');
-var signinRouter = require('./routes/signin');
+var {router: notesRouter} = require('./routes/notes');
+var {router: signupRouter} = require('./routes/signup');
+var {router: signinRouter} = require('./routes/signin');
 
 var app = express();
 
@@ -22,8 +20,6 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/', indexRouter);
-app.use('/users', usersRouter);
 app.use('/notes', notesRouter);
 app.use('/signup', signupRouter);
 app.use('/signin', signinRouter);
