+++ routes/validStrings.js
@@ -1,0 +1,13 @@
+function isUsernameValid(str){
+    if(typeof(str)!== 'string'){
+        return false;
+    }
+    for(var i=0;i<str.length;i++){
+        if(str.charCodeAt(i)>122 || str.charCodeAt(i)<97){
+            return false;
+        }
+    }
+    return true;
+}
+
+module.exports = {isUsernameValid};
